requirements.txt
absl-py==1.2.0
astunparse==1.6.3
attrs==22.1.0
cachetools==5.2.0
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
contourpy==1.0.5
cryptography==38.0.1
cycler==0.11.0
flatbuffers==2.0.7
fonttools==4.37.2
gast==0.4.0
google-api-core==2.10.1
google-api-python-client==2.61.0
google-auth==2.11.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.56.4
grpcio==1.49.0
h5py==3.7.0
httplib2==0.20.4
idna==3.4
joblib==1.2.0
keras==2.10.0
keras-nightly==2.11.0.dev2022091707
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.0
mediapipe==0.8.11
numpy==1.23.3
oauth2client==4.1.3
oauthlib==3.2.1
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
Pillow==9.2.0
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDrive2==1.14.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
qrcode==7.3.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.1
six==1.16.0
tb-nightly==2.11.0a20220917
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
tf-estimator-nightly==2.11.0.dev2022091708
tf-nightly==2.11.0.dev20220917
threadpoolctl==3.1.0
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
